# swagger/json/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a Restful API server by Renj.io. Generated by [http://swagger.io](http://swagger.io). JJgo is written in Go and serving all the JJ's APPs. Learn more at [JJService](http://app.renj.io)",
    "version": "1.0.0",
    "title": "JJGo Rest API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "mail@renj.io"
    }
  },
  "host": "api.renj.io",
  "basePath": "/rest",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/jjmail/sub_blog": {
      "put": {
        "tags": [
          "jjmail"
        ],
        "summary": "jjmail subscribe",
        "description": "jjmail订阅博客文章邮件",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "根据收件邮箱地址和所选服务名添加任务",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "mail_address": {
                  "type": "string",
                  "description": "收件人地址"
                },
                "service": {
                  "type": "string",
                  "description": "所选服务模块"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success response",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/jjmail/unsub_blog": {
      "delete": {
        "tags": [
          "jjmail"
        ],
        "summary": "jjmail unsubscribe",
        "description": "jjmail取消订阅博客文章",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "根据收件邮箱地址和所选服务名添加任务",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "mail_address": {
                  "type": "string",
                  "description": "收件人地址"
                },
                "service": {
                  "type": "string",
                  "description": "所选服务模块"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success response",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/jjmail/sub_mgek": {
      "put": {
        "tags": [
          "jjmail"
        ],
        "summary": "jjmail subscribe",
        "description": "jjmail订阅mgek站点邮件",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "根据收件邮箱地址和所选服务名添加任务",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "mail_address": {
                  "type": "string",
                  "description": "收件人地址"
                },
                "service": {
                  "type": "string",
                  "description": "所选服务模块"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success response",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/jjmail/unsub_mgek": {
      "delete": {
        "tags": [
          "jjmail"
        ],
        "summary": "jjmail unsubscribe",
        "description": "jjmail取消订阅mgek站点",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "根据收件邮箱地址和所选服务名添加任务",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "mail_address": {
                  "type": "string",
                  "description": "收件人地址"
                },
                "service": {
                  "type": "string",
                  "description": "所选服务模块"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success response",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "default": 200,
          "format": "int32",
          "description": "HTTP status code"
        },
        "time": {
          "type": "string",
          "default": "1000ms",
          "description": "time latency"
        },
        "mes": {
          "type": "string",
          "description": "response message"
        },
        "data": {
          "type": "object",
          "description": "result data"
        }
      }
    },
...
  },
...
}